Deleted or moved some unessisary stuff
